% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/catalog.R
\name{as.data.frame.dss_catalog}
\alias{as.data.frame.dss_catalog}
\title{Convert a dss_catalog to a data.frame for inspection.}
\usage{
\method{as.data.frame}{dss_catalog}(x, row.names = NULL, optional = FALSE, ...)
}
\description{
Convert a dss_catalog to a data.frame for inspection.
}
