% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dateconverter.R
\name{datetime_from_dss_datetime_string}
\alias{datetime_from_dss_datetime_string}
\title{Parse a DSS-style "DDmonYYYY HH:MM[:SS]" string with the 24:00 rule.}
\usage{
datetime_from_dss_datetime_string(s)
}
\description{
Parse a DSS-style "DDmonYYYY HH:MM[:SS]" string with the 24:00 rule.
}
