% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dateconverter.R
\name{dss_datetime_strings_from_datetime}
\alias{dss_datetime_strings_from_datetime}
\title{Convert POSIXct to DSS-style "DDmonYYYY HH:MM:SS" with the 24:00 midnight rule.}
\usage{
dss_datetime_strings_from_datetime(dt)
}
\value{
list(date, time).
}
\description{
Convert POSIXct to DSS-style "DDmonYYYY HH:MM:SS" with the 24:00 midnight rule.
}
