% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dsspath.R
\name{format_dss_path}
\alias{format_dss_path}
\title{Format a dss_path back to its string form.}
\usage{
format_dss_path(p)
}
\arguments{
\item{p}{dss_path object.}
}
\description{
Format a dss_path back to its string form.
}
