% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hecdss.R
\name{HecDss}
\alias{HecDss}
\title{HecDss: open and interact with a DSS file.}
\description{
Phase 1 supports regular and irregular time series records. Other record
types (paired data, gridded data, etc.) will be added in later phases.
}
\examples{
\dontrun{
  dss <- HecDss$new("examples-all-data-types.dss")
  cat <- dss$get_catalog()
  ts  <- dss$get("/regular-time-series/GAPT/FLOW/01Sep2021/6Hour/forecast1/")
  plot(ts)
  dss$close()
}
}
\section{Methods}{
\subsection{Public methods}{
  \itemize{
    \item \href{#method-HecDss-initialize}{\code{HecDss$new()}}
    \item \href{#method-HecDss-close}{\code{HecDss$close()}}
    \item \href{#method-HecDss-get_version}{\code{HecDss$get_version()}}
    \item \href{#method-HecDss-record_count}{\code{HecDss$record_count()}}
    \item \href{#method-HecDss-get_catalog}{\code{HecDss$get_catalog()}}
    \item \href{#method-HecDss-get_record_type}{\code{HecDss$get_record_type()}}
    \item \href{#method-HecDss-get}{\code{HecDss$get()}}
    \item \href{#method-HecDss-put}{\code{HecDss$put()}}
    \item \href{#method-HecDss-delete}{\code{HecDss$delete()}}
    \item \href{#method-HecDss-set_debug_level}{\code{HecDss$set_debug_level()}}
    \item \href{#method-HecDss-is_open}{\code{HecDss$is_open()}}
    \item \href{#method-HecDss-clone}{\code{HecDss$clone()}}
  }
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-HecDss-initialize"></a>}}
\if{latex}{\out{\hypertarget{method-HecDss-initialize}{}}}
\subsection{\code{HecDss$new()}}{
  \subsection{Usage}{
    \if{html}{\out{<div class="r">}}
    \preformatted{HecDss$new(filename)}
    \if{html}{\out{</div>}}
  }
}

\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-HecDss-close"></a>}}
\if{latex}{\out{\hypertarget{method-HecDss-close}{}}}
\subsection{\code{HecDss$close()}}{
  \subsection{Usage}{
    \if{html}{\out{<div class="r">}}
    \preformatted{HecDss$close()}
    \if{html}{\out{</div>}}
  }
}

\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-HecDss-get_version"></a>}}
\if{latex}{\out{\hypertarget{method-HecDss-get_version}{}}}
\subsection{\code{HecDss$get_version()}}{
  \subsection{Usage}{
    \if{html}{\out{<div class="r">}}
    \preformatted{HecDss$get_version()}
    \if{html}{\out{</div>}}
  }
}

\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-HecDss-record_count"></a>}}
\if{latex}{\out{\hypertarget{method-HecDss-record_count}{}}}
\subsection{\code{HecDss$record_count()}}{
  \subsection{Usage}{
    \if{html}{\out{<div class="r">}}
    \preformatted{HecDss$record_count()}
    \if{html}{\out{</div>}}
  }
}

\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-HecDss-get_catalog"></a>}}
\if{latex}{\out{\hypertarget{method-HecDss-get_catalog}{}}}
\subsection{\code{HecDss$get_catalog()}}{
  \subsection{Usage}{
    \if{html}{\out{<div class="r">}}
    \preformatted{HecDss$get_catalog(filter = "", refresh = FALSE)}
    \if{html}{\out{</div>}}
  }
}

\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-HecDss-get_record_type"></a>}}
\if{latex}{\out{\hypertarget{method-HecDss-get_record_type}{}}}
\subsection{\code{HecDss$get_record_type()}}{
  \subsection{Usage}{
    \if{html}{\out{<div class="r">}}
    \preformatted{HecDss$get_record_type(pathname)}
    \if{html}{\out{</div>}}
  }
}

\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-HecDss-get"></a>}}
\if{latex}{\out{\hypertarget{method-HecDss-get}{}}}
\subsection{\code{HecDss$get()}}{
  \subsection{Usage}{
    \if{html}{\out{<div class="r">}}
    \preformatted{HecDss$get(
  pathname,
  start_datetime = NULL,
  end_datetime = NULL,
  trim = FALSE,
  tz = "UTC"
)}
    \if{html}{\out{</div>}}
  }
}

\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-HecDss-put"></a>}}
\if{latex}{\out{\hypertarget{method-HecDss-put}{}}}
\subsection{\code{HecDss$put()}}{
  \subsection{Usage}{
    \if{html}{\out{<div class="r">}}
    \preformatted{HecDss$put(
  ts,
  pathname = NULL,
  type = c("regular", "irregular"),
  time_granularity_seconds = 60L
)}
    \if{html}{\out{</div>}}
  }
}

\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-HecDss-delete"></a>}}
\if{latex}{\out{\hypertarget{method-HecDss-delete}{}}}
\subsection{\code{HecDss$delete()}}{
  \subsection{Usage}{
    \if{html}{\out{<div class="r">}}
    \preformatted{HecDss$delete(pathname)}
    \if{html}{\out{</div>}}
  }
}

\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-HecDss-set_debug_level"></a>}}
\if{latex}{\out{\hypertarget{method-HecDss-set_debug_level}{}}}
\subsection{\code{HecDss$set_debug_level()}}{
  \subsection{Usage}{
    \if{html}{\out{<div class="r">}}
    \preformatted{HecDss$set_debug_level(level)}
    \if{html}{\out{</div>}}
  }
}

\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-HecDss-is_open"></a>}}
\if{latex}{\out{\hypertarget{method-HecDss-is_open}{}}}
\subsection{\code{HecDss$is_open()}}{
  \subsection{Usage}{
    \if{html}{\out{<div class="r">}}
    \preformatted{HecDss$is_open()}
    \if{html}{\out{</div>}}
  }
}

\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-HecDss-clone"></a>}}
\if{latex}{\out{\hypertarget{method-HecDss-clone}{}}}
\subsection{\code{HecDss$clone()}}{
  The objects of this class are cloneable with this method.
  \subsection{Usage}{
    \if{html}{\out{<div class="r">}}
    \preformatted{HecDss$clone(deep = FALSE)}
    \if{html}{\out{</div>}}
  }
  \subsection{Arguments}{
    \if{html}{\out{<div class="arguments">}}
    \describe{
      \item{\code{deep}}{Whether to make a deep clone.}
    }
    \if{html}{\out{</div>}}
  }
}

}
