% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hecdssr-package.R
\docType{package}
\name{hecdssr-package}
\alias{hecdssr}
\alias{hecdssr-package}
\title{hecdssr: R Interface to HEC-DSS}
\description{
R bindings for the HEC-DSS C library, mirroring the architecture of
\href{https://github.com/HydrologicEngineeringCenter/hec-dss-python}{hec-dss-python}:
a thin Rcpp shim over the hecdss C API with all higher-level logic in R.
}
\details{
The package is named `hecdssr` (not `hecdss`) to avoid a Windows
DLL-name collision with the underlying HEC C library `hecdss.dll`.
}
\author{
\strong{Maintainer}: HEC Developer \email{hecdss@example.com}

Authors:
\itemize{
  \item HEC Developer \email{hecdss@example.com}
}

}
\keyword{internal}
