% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dsspath.R
\name{parse_dss_path}
\alias{parse_dss_path}
\title{Parse a DSS pathname.}
\usage{
parse_dss_path(path, rec_type = "Unknown")
}
\arguments{
\item{path}{Raw DSS pathname starting and ending with "/".}

\item{rec_type}{Optional record type name (see RECORD_TYPE).}
}
\value{
An object of class "dss_path" with fields A,B,C,D,E,F,rec_type.
}
\description{
Parse a DSS pathname.
}
