% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dsspath.R
\name{path_without_date}
\alias{path_without_date}
\title{Return the path with the D (date) part blanked out.}
\usage{
path_without_date(p)
}
\arguments{
\item{p}{dss_path object.}
}
\description{
Return the path with the D (date) part blanked out.
}
