% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/record_type.R
\name{record_type_from_int}
\alias{record_type_from_int}
\title{Convert a raw integer record type from the DSS API to a symbolic name.}
\usage{
record_type_from_int(rec_type)
}
\arguments{
\item{rec_type}{integer raw record type from hec_dss_recordType / catalog.}
}
\value{
character: one of names(RECORD_TYPE).
}
\description{
Mirrors RecordType.RecordTypeFromInt in hec-dss-python.
}
