% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/irregular_timeseries.R
\name{store_irregular_timeseries}
\alias{store_irregular_timeseries}
\title{Store an xts as an irregular DSS time series.}
\usage{
store_irregular_timeseries(
  dss,
  ts,
  pathname = NULL,
  time_granularity_seconds = 60L
)
}
\arguments{
\item{dss}{DssPtr from native_open / open_dss.}

\item{ts}{xts. Index is POSIXct; first column is the value.}

\item{pathname}{Optional pathname override.}

\item{time_granularity_seconds}{1, 60, 3600, or 86400. Default 60.}
}
\description{
Store an xts as an irregular DSS time series.
}
